LPS-65385 Move DestinationNames to com.liferay.portal.security.ldap.internal.constants
